// frontend/src/app/features/feature-issues/models/IssueStatusRequest.ts
export enum IssueStatusRequest {
    assignToDeveloper = "assignToDeveloper",
    closeByDeveloper = "closeByDeveloper",
    verify = "verify",
    retest = "retest",
    sendToRetest = "sendToRetest",
    fix = "fix",
    duplicate = "duplicate",
    reject = "reject",
    defer = "defer",
    notABug = "notABug",
    open = "open",
    reopen = "reopen"
}